+++ src/helpers.h
@@ -13,7 +13,7 @@ union INDEX_TYPE_t
 {
   //var[MAXVAR] gives the bits that represent the left and right children
   unsigned short var[MAXVAR+1];   
-  unsigned long long index;       // probably not used
+  unsigned long index;       // probably not used
 };
 typedef INDEX_TYPE_t INDEX_TYPE;
 
@@ -28,11 +28,11 @@ struct side_type
 struct cube_type
 {
   std::vector<side_type> sides;
-  ushort level;
+  unsigned short level;
   double alt_prob;
   vec effect_size;         
   int direction;
-  ushort node_idx;  
+  unsigned short node_idx;  
   arma::Col< unsigned > data_points;
 };
 
@@ -42,27 +42,27 @@ typedef std::vector<cube_type> result_cubes_type;
 
 double log_exp_x_plus_exp_y(double x, double y);
 
-unsigned long long pow2(int k); 
+unsigned long pow2(int k); 
 
-uint convert_to_inverse_base_2(double x, int k);
+unsigned int convert_to_inverse_base_2(double x, int k);
 
-unsigned long long int Choose(int n, int k);
+unsigned long int Choose(int n, int k);
 
 INDEX_TYPE init_index(int level); 
 
-unsigned long long int get_node_index(INDEX_TYPE& I,
+unsigned long int get_node_index(INDEX_TYPE& I,
                                       int level, 
                                       int dim);
                                       
 std::pair<bool, INDEX_TYPE> make_parent_index( INDEX_TYPE& I, 
                                                 unsigned short part_dim, 
                                                 int level, 
-                                                ushort which );
+                                                unsigned short which );
 
 INDEX_TYPE make_child_index(  INDEX_TYPE& I, 
                               unsigned short part_dim, 
                               int level, 
-                              ushort which); 
+                              unsigned short which); 
                               
 int sum_elem(int * my_array , int num_elem);
